+++ src/analyzer.ts
@@ -1,13 +1,13 @@
 import { findFilesWithImports } from "./traversar";
 import { createDisallowedImportViolation, Violation } from "./violations";
 import { getDemand } from "./store";
-import { ImportConfig } from "./config";
 import { getFiles } from "./utils";
+import { TImportConfig } from "./config/types";
 
 export const analyze = async (
   paths: string[],
-  disallowedImports: ImportConfig[],
-  allowedImports: ImportConfig[],
+  disallowedImports: TImportConfig[],
+  allowedImports: TImportConfig[],
 ): Promise<Violation[]> =>
   paths.flatMap(path => {
     const demand = getDemand(path);
